Er_mostro/llama3_rag_setup/eda_engine/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
python-dotenv
loguru
streamlit
requests